tcc/pyproject.toml
[project]
name = "tcc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=4.3.0",
    "fastapi[standard]>=0.115.12",
    "psycopg2-binary>=2.9.10",
    "py-solc-x>=2.0.3",
    "pyjwt>=2.10.1",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "ruff>=0.11.5",
    "web3>=7.11.0",
]

